feat/async snapshot (#287)
* (feat) snapshot backup for MemoryKVStore

* (feat) snapshot backup for MemoryKVStore

* (feat) snapshot backup for MemoryKVStore

* (feat) async snapshot for rocksdb

* (feat) typo

* (feat) add isAvailable() method

* (feat) make option `fastSnapshot` default true

* (feat) code format

* (feat) code format

* (feat) async snapshot option

* feat/checksum snapshot (#289)

* (feat) rheakv snapshot checksum

* (feat) checksum test

* (feat) code format

* (feat) add mor unit test

* (feat) change checksum impl to CRC64

* (feat) comment

* (feat) avoid memory copy of checksum in LogEntry (#293)

+++ jraft-rheakv/rheakv-core/src/main/java/com/alipay/sofa/jraft/rhea/storage/RocksRawKVStore.java
@@ -25,6 +25,8 @@ import java.util.Comparator;
 import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.locks.Lock;
@@ -82,6 +84,7 @@ import com.alipay.sofa.jraft.util.BytesUtil;
 import com.alipay.sofa.jraft.util.Requires;
 import com.alipay.sofa.jraft.util.StorageOptionsFactory;
 import com.alipay.sofa.jraft.util.SystemPropertyUtil;
+import com.alipay.sofa.jraft.util.concurrent.AdjustableSemaphore;
 import com.codahale.metrics.Timer;
 
 /**
@@ -102,6 +105,7 @@ public class RocksRawKVStore extends BatchRawKVStore<RocksDBOptions> {
     public static final int                    MAX_BATCH_WRITE_SIZE = SystemPropertyUtil.getInt(
                                                                         "rhea.rocksdb.user.max_batch_write_size", 128);
 
+    private final AdjustableSemaphore          shutdownLock         = new AdjustableSemaphore();
     private final ReadWriteLock                readWriteLock        = new ReentrantReadWriteLock();
 
     private final AtomicLong                   databaseVersion      = new AtomicLong(0);
@@ -161,6 +165,7 @@ public class RocksRawKVStore extends BatchRawKVStore<RocksDBOptions> {
             // to reply to the data is the correct behavior.
             destroyRocksDB(opts);
             openRocksDB(opts);
+            this.shutdownLock.setMaxPermits(1);
             LOG.info("[RocksRawKVStore] start successfully, options: {}.", opts);
             return true;
         } catch (final Exception e) {
@@ -179,6 +184,7 @@ public class RocksRawKVStore extends BatchRawKVStore<RocksDBOptions> {
             if (this.db == null) {
                 return;
             }
+            this.shutdownLock.setMaxPermits(0);
             closeRocksDB();
             if (this.defaultHandle != null) {
                 this.defaultHandle.close();
@@ -1210,55 +1216,91 @@ public class RocksRawKVStore extends BatchRawKVStore<RocksDBOptions> {
         return Requires.requireNonNull(this.opts, "opts").isFastSnapshot();
     }
 
-    void createSstFiles(final EnumMap<SstColumnFamily, File> sstFileTable, final byte[] startKey, final byte[] endKey) {
+    boolean isAsyncSnapshot() {
+        return Requires.requireNonNull(this.opts, "opts").isAsyncSnapshot();
+    }
+
+    CompletableFuture<Void> createSstFiles(final EnumMap<SstColumnFamily, File> sstFileTable, final byte[] startKey,
+                                           final byte[] endKey, final ExecutorService executor) {
+        final Snapshot snapshot;
+        final CompletableFuture<Void> sstFuture = new CompletableFuture<>();
+        final Lock readLock = this.readWriteLock.readLock();
+        readLock.lock();
+        try {
+            snapshot = this.db.getSnapshot();
+            if (!isAsyncSnapshot()) {
+                doCreateSstFiles(snapshot, sstFileTable, startKey, endKey, sstFuture);
+                return sstFuture;
+            }
+        } finally {
+            readLock.unlock();
+        }
+
+        // async snapshot
+        executor.execute(() -> doCreateSstFiles(snapshot, sstFileTable, startKey, endKey, sstFuture));
+        return sstFuture;
+    }
+
+    void doCreateSstFiles(final Snapshot snapshot, final EnumMap<SstColumnFamily, File> sstFileTable,
+                          final byte[] startKey, final byte[] endKey, final CompletableFuture<Void> future) {
         final Timer.Context timeCtx = getTimeContext("CREATE_SST_FILE");
         final Lock readLock = this.readWriteLock.readLock();
         readLock.lock();
-        final Snapshot snapshot = this.db.getSnapshot();
-        try (final ReadOptions readOptions = new ReadOptions();
-                final EnvOptions envOptions = new EnvOptions();
-                final Options options = new Options().setMergeOperator(new StringAppendOperator())) {
-            readOptions.setSnapshot(snapshot);
-            for (final Map.Entry<SstColumnFamily, File> entry : sstFileTable.entrySet()) {
-                final SstColumnFamily sstColumnFamily = entry.getKey();
-                final File sstFile = entry.getValue();
-                final ColumnFamilyHandle columnFamilyHandle = findColumnFamilyHandle(sstColumnFamily);
-                try (final RocksIterator it = this.db.newIterator(columnFamilyHandle, readOptions);
-                        final SstFileWriter sstFileWriter = new SstFileWriter(envOptions, options)) {
-                    if (startKey == null) {
-                        it.seekToFirst();
-                    } else {
-                        it.seek(startKey);
-                    }
-                    sstFileWriter.open(sstFile.getAbsolutePath());
-                    long count = 0;
-                    for (;;) {
-                        if (!it.isValid()) {
-                            break;
+        try {
+            if (!this.shutdownLock.isAvailable()) {
+                // KV store has shutdown, we do not release rocksdb's snapshot
+                future.completeExceptionally(new StorageException("KV store has shutdown."));
+                return;
+            }
+            try (final ReadOptions readOptions = new ReadOptions();
+                    final EnvOptions envOptions = new EnvOptions();
+                    final Options options = new Options().setMergeOperator(new StringAppendOperator())) {
+                readOptions.setSnapshot(snapshot);
+                for (final Map.Entry<SstColumnFamily, File> entry : sstFileTable.entrySet()) {
+                    final SstColumnFamily sstColumnFamily = entry.getKey();
+                    final File sstFile = entry.getValue();
+                    final ColumnFamilyHandle columnFamilyHandle = findColumnFamilyHandle(sstColumnFamily);
+                    try (final RocksIterator it = this.db.newIterator(columnFamilyHandle, readOptions);
+                            final SstFileWriter sstFileWriter = new SstFileWriter(envOptions, options)) {
+                        if (startKey == null) {
+                            it.seekToFirst();
+                        } else {
+                            it.seek(startKey);
                         }
-                        final byte[] key = it.key();
-                        if (endKey != null && BytesUtil.compare(key, endKey) >= 0) {
-                            break;
+                        sstFileWriter.open(sstFile.getAbsolutePath());
+                        long count = 0;
+                        for (;;) {
+                            if (!it.isValid()) {
+                                break;
+                            }
+                            final byte[] key = it.key();
+                            if (endKey != null && BytesUtil.compare(key, endKey) >= 0) {
+                                break;
+                            }
+                            sstFileWriter.put(key, it.value());
+                            ++count;
+                            it.next();
                         }
-                        sstFileWriter.put(key, it.value());
-                        ++count;
-                        it.next();
+                        if (count == 0) {
+                            sstFileWriter.close();
+                        } else {
+                            sstFileWriter.finish();
+                        }
+                        LOG.info("Finish sst file {} with {} keys.", sstFile, count);
+                    } catch (final RocksDBException e) {
+                        throw new StorageException("Fail to create sst file at path: " + sstFile, e);
                     }
-                    if (count == 0) {
-                        sstFileWriter.close();
-                    } else {
-                        sstFileWriter.finish();
-                    }
-                    LOG.info("Finish sst file {} with {} keys.", sstFile, count);
-                } catch (final RocksDBException e) {
-                    throw new StorageException("Fail to create sst file at path: " + sstFile, e);
                 }
+                future.complete(null);
+            } catch (final Throwable t) {
+                future.completeExceptionally(t);
+            } finally {
+                // Nothing to release, rocksDB never own the pointer for a snapshot.
+                snapshot.close();
+                // The pointer to the snapshot is released by the database instance.
+                this.db.releaseSnapshot(snapshot);
             }
         } finally {
-            // Nothing to release, rocksDB never own the pointer for a snapshot.
-            snapshot.close();
-            // The pointer to the snapshot is released by the database instance.
-            this.db.releaseSnapshot(snapshot);
             readLock.unlock();
             timeCtx.stop();
         }
@@ -1388,7 +1430,7 @@ public class RocksRawKVStore extends BatchRawKVStore<RocksDBOptions> {
         }
     }
 
-    void writeSstSnapshot(final String snapshotPath, final Region region) {
+    CompletableFuture<Void> writeSstSnapshot(final String snapshotPath, final Region region, final ExecutorService executor) {
         final Timer.Context timeCtx = getTimeContext("WRITE_SST_SNAPSHOT");
         final Lock readLock = this.readWriteLock.readLock();
         readLock.lock();
@@ -1399,12 +1441,26 @@ public class RocksRawKVStore extends BatchRawKVStore<RocksDBOptions> {
             FileUtils.forceMkdir(tempFile);
 
             final EnumMap<SstColumnFamily, File> sstFileTable = getSstFileTable(tempPath);
-            createSstFiles(sstFileTable, region.getStartKey(), region.getEndKey());
-            final File snapshotFile = new File(snapshotPath);
-            FileUtils.deleteDirectory(snapshotFile);
-            if (!tempFile.renameTo(snapshotFile)) {
-                throw new StorageException("Fail to rename [" + tempPath + "] to [" + snapshotPath + "].");
-            }
+            final CompletableFuture<Void> snapshotFuture = new CompletableFuture<>();
+            final CompletableFuture<Void> sstFuture = createSstFiles(sstFileTable, region.getStartKey(),
+                region.getEndKey(), executor);
+            sstFuture.whenComplete((aVoid, throwable) -> {
+                if (throwable == null) {
+                    try {
+                        final File snapshotFile = new File(snapshotPath);
+                        FileUtils.deleteDirectory(snapshotFile);
+                        if (!tempFile.renameTo(snapshotFile)) {
+                            throw new StorageException("Fail to rename [" + tempPath + "] to [" + snapshotPath + "].");
+                        }
+                        snapshotFuture.complete(null);
+                    } catch (final Throwable t) {
+                        snapshotFuture.completeExceptionally(t);
+                    }
+                } else {
+                    snapshotFuture.completeExceptionally(throwable);
+                }
+            });
+            return snapshotFuture;
         } catch (final Exception e) {
             throw new StorageException("Fail to do read sst snapshot at path: " + snapshotPath, e);
         } finally {
